PHRAS-3092 #comment merge api setmetadatas v3

+++ lib/classes/record/adapter.php
@@ -1054,10 +1054,13 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
             }
         }
 
-        if (trim($params['meta_id']) !== '') {
-            $tmp_val = trim($params['value']);
+        $tmp_val = trim($params['value']);
 
-            $caption_field_value = $caption_field->get_value($params['meta_id']);
+        if (trim($params['meta_id']) !== '') {
+
+            if(is_null($caption_field_value = $caption_field->get_value($params['meta_id']))) {
+                return $this;
+            }
 
             if ($tmp_val === '') {
                 $caption_field_value->delete();
@@ -1068,8 +1071,11 @@ class record_adapter implements RecordInterface, cache_cacheableInterface
                     $caption_field_value->setVocab($vocab, $vocab_id);
                 }
             }
-        } else {
-            $caption_field_value = caption_Field_Value::create($this->app, $databox_field, $this, $params['value'], $vocab, $vocab_id);
+        }
+        else {
+            if($tmp_val !== '') {
+                caption_Field_Value::create($this->app, $databox_field, $this, $params['value'], $vocab, $vocab_id);
+            }
         }
 
         return $this;
